requirements.txt
asgiref==3.2.7
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.10.2
install==1.3.1
mysqlclient==1.4.6
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
